Rewrite log handling to isolate log collection and rendering
