+++ DTXMania/UI/Inspector/LogWindow.cs
@@ -1,125 +1,23 @@
 ﻿using System.Diagnostics;
-using System.Text;
 using System.Numerics;
+using DTXMania.Core;
+using DTXMania.Core.Framework;
 using Hexa.NET.ImGui;
 
 namespace DTXMania.UI.Inspector;
 
-public class LogWindow : TraceListener
+public class LogWindow
 {
-    private const int MaxLogLines = 200_000;
-
-    private readonly RingBuffer<LogLine> logLines = new(MaxLogLines);
-    private readonly object logLock = new();
-
     private bool autoScroll = true;
-    private readonly StringBuilder messageBuffer = new();
+    public RuntimeLogListener? logSource;
 
-    private struct LogLine
-    {
-        public string Text;
-        public TraceEventType Level;
-    }
-
-    public override void Write(string? message)
-    {
-        if (message == null) return;
-
-        lock (logLock)
-        {
-            messageBuffer.Append(message);
-        }
-    }
-
-    public override void WriteLine(string? message)
-    {
-        if (message == null) return;
-
-        lock (logLock)
-        {
-            messageBuffer.AppendLine(message);
-            string fullMessage = messageBuffer.ToString().TrimEnd('\n', '\r');
-            messageBuffer.Clear();
-
-            AppendMessageLocked(fullMessage, TraceEventType.Information);
-        }
-    }
-
-    public override void TraceEvent(TraceEventCache? eventCache, string? source, TraceEventType eventType, int id, string? message)
-    {
-        if (message == null) return;
-
-        lock (logLock)
-        {
-            AppendMessageLocked(message, eventType);
-        }
-    }
-
-    public override void TraceEvent(TraceEventCache? eventCache, string? source, TraceEventType eventType, int id, string? format, params object?[]? args)
-    {
-        if (format == null) return;
-
-        lock (logLock)
-        {
-            string message = args == null ? format : string.Format(format, args);
-            AppendMessageLocked(message, eventType);
-        }
-    }
-
-    public override void TraceData(TraceEventCache? eventCache, string? source, TraceEventType eventType, int id, object? data)
-    {
-        if (data == null) return;
-
-        lock (logLock)
-        {
-            AppendMessageLocked(data.ToString() ?? string.Empty, eventType);
-        }
-    }
-
-    public override void TraceData(TraceEventCache? eventCache, string? source, TraceEventType eventType, int id, params object?[]? data)
-    {
-        if (data == null) return;
-
-        lock (logLock)
-        {
-            StringBuilder sb = new();
-            for (int i = 0; i < data.Length; i++)
-            {
-                if (i > 0) sb.Append(", ");
-                sb.Append(data[i]?.ToString());
-            }
-            AppendMessageLocked(sb.ToString(), eventType);
-        }
-    }
-
-    private void AppendMessageLocked(string message, TraceEventType level)
-    {
-        string prefix = $"[{DateTime.Now:HH:mm:ss.fff}] [{level}] ";
-
-        if (string.IsNullOrEmpty(message))
-        {
-            logLines.Add(new LogLine { Text = prefix, Level = level });
-            return;
-        }
-
-        string[] lines = message.Replace("\r\n", "\n").Replace('\r', '\n').Split('\n');
-        foreach (string line in lines)
-        {
-            logLines.Add(new LogLine { Text = prefix + line, Level = level });
-        }
-    }
-    
     public void DrawWindow()
     {
         ImGui.Begin("Log Window", ImGuiWindowFlags.NoFocusOnAppearing);
 
         if (ImGui.Button("Clear"))
         {
-            lock (logLock)
-            {
-                logLines.Clear();
-                messageBuffer.Clear();
-            }
+            logSource?.Clear();
         }
 
         ImGui.SameLine();
@@ -130,28 +28,31 @@ public class LogWindow : TraceListener
         // Determine bottom state before adding this frame's content.
         bool wasNearBottom = ImGui.GetScrollY() >= ImGui.GetScrollMaxY() - 10f;
 
-        lock (logLock)
+        if (logSource != null)
         {
-            int count = logLines.Count;
-            if (count > 0)
+            lock (logSource.logLock)
             {
-                var clipper = new ImGuiListClipper();
-                clipper.Begin(count);
-
-                while (clipper.Step())
+                int count = logSource.logLines.Count;
+                if (count > 0)
                 {
-                    for (int i = clipper.DisplayStart; i < clipper.DisplayEnd; i++)
+                    ImGuiListClipper clipper = new();
+                    clipper.Begin(count);
+
+                    while (clipper.Step())
                     {
-                        LogLine line = logLines[i];
-                        Vector4 color = GetColorForLevel(line.Level);
+                        for (int i = clipper.DisplayStart; i < clipper.DisplayEnd; i++)
+                        {
+                            RuntimeLogListener.LogLine line = logSource.logLines[i];
+                            Vector4 color = GetColorForLevel(line.Level);
 
-                        ImGui.PushStyleColor(ImGuiCol.Text, color);
-                        ImGui.TextUnformatted(line.Text);
-                        ImGui.PopStyleColor();
+                            ImGui.PushStyleColor(ImGuiCol.Text, color);
+                            ImGui.TextUnformatted(line.Text);
+                            ImGui.PopStyleColor();
+                        }
                     }
-                }
 
-                clipper.End();
+                    clipper.End();
+                }
             }
         }
 
@@ -164,58 +65,7 @@ public class LogWindow : TraceListener
         ImGui.End();
     }
 
-    private sealed class RingBuffer<T>
-    {
-        private readonly T[] buffer;
-        private int start;
-        private int count;
-
-        public RingBuffer(int capacity)
-        {
-            if (capacity <= 0)
-            {
-                throw new ArgumentOutOfRangeException(nameof(capacity));
-            }
-
-            buffer = new T[capacity];
-        }
-
-        public int Count => count;
-
-        public T this[int index]
-        {
-            get
-            {
-                if ((uint)index >= (uint)count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index));
-                }
-
-                return buffer[(start + index) % buffer.Length];
-            }
-        }
-
-        public void Add(T item)
-        {
-            if (count < buffer.Length)
-            {
-                buffer[(start + count) % buffer.Length] = item;
-                count++;
-                return;
-            }
-
-            buffer[start] = item;
-            start = (start + 1) % buffer.Length;
-        }
-
-        public void Clear()
-        {
-            start = 0;
-            count = 0;
-        }
-    }
-
-    private static Vector4 GetColorForLevel(TraceEventType level) => level switch
+    public static Vector4 GetColorForLevel(TraceEventType level) => level switch
     {
         TraceEventType.Critical => new Vector4(1f, 0.2f, 0.2f, 1f),
         TraceEventType.Error => new Vector4(1f, 0.4f, 0.4f, 1f),
